sas-clinical-macros/requirements.txt
# Python dependencies for SAS Clinical Macros CI/CD

# Core dependencies
pyyaml>=6.0
requests>=2.28.0
numpy>=1.24.0

# Documentation and search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.3
flask>=2.3.0
flask-cors>=4.0.0

# Testing and validation
pytest>=7.3.0
pytest-html>=3.2.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
colorama>=0.4.6
tabulate>=0.9.0

# Development tools
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.0